+++ src/components/autoComplete.js
@@ -68,6 +68,7 @@
     const { kind, values: listValues } = getProperty(property) || {};
     const [currentValue, setCurrentValue] = useState(useText(defaultValue));
     const [currentLabel, setCurrentLabel] = useState('');
+    const mounted = useRef(false);
     const labelText = useText(label);
 
     const textFieldProps = {
@@ -151,7 +152,7 @@
           }, {})
         : {};
 
-    const { loading, error: err, data, refetch } =
+    const { loading, error: err, data: { results } = {}, refetch } =
       model &&
       useAllQuery(model, {
         ...useFilter,
@@ -160,13 +161,39 @@
         variables: {
           ...(orderBy ? { sort: { relation: sort } } : {}),
         },
+        onCompleted(res) {
+          const hasResult = res && res.result && res.result.length > 0;
+          if (hasResult) {
+            B.triggerEvent('onSuccess', res.results);
+          } else {
+            B.triggerEvent('onNoResults');
+          }
+        },
+        onError(resp) {
+          if (!displayError) {
+            B.triggerEvent('onError', resp);
+          }
+        },
       });
 
     useEffect(() => {
-      if (!isDev && data) {
+      mounted.current = true;
+      return () => {
+        mounted.current = false;
+      };
+    }, []);
+
+    useEffect(() => {
+      if (!isDev && results) {
         resetFilter();
       }
-    }, [data]);
+    }, [results]);
+
+    useEffect(() => {
+      if (mounted.current && loading) {
+        B.triggerEvent('onLoad', loading);
+      }
+    }, [loading]);
 
     useEffect(() => {
       if (isDev) {
@@ -174,7 +201,10 @@
       }
     }, [isDev, defaultValue]);
 
-    B.defineFunction('Clear', () => setCurrentValue(null));
+    B.defineFunction('Clear', () => {
+      setCurrentValue(multiple ? [] : null);
+      setSearchParam('');
+    });
     B.defineFunction('Refetch', () => refetch());
 
     useEffect(() => {
@@ -185,34 +215,6 @@
         clearTimeout(handler);
       };
     }, [searchParam]);
-
-    const mounted = useRef(false);
-
-    useEffect(() => {
-      mounted.current = true;
-      return () => {
-        mounted.current = false;
-      };
-    }, []);
-
-    useEffect(() => {
-      if (mounted.current && loading) {
-        B.triggerEvent('onLoad', loading);
-      }
-    }, [loading]);
-
-    if (err && !displayError) {
-      B.triggerEvent('onError', err);
-    }
-
-    const { results } = data || {};
-    if (results) {
-      if (results.length > 0) {
-        B.triggerEvent('onSuccess', results);
-      } else {
-        B.triggerEvent('onNoResults');
-      }
-    }
 
     const onChange = (_, newValue) => {
       if (!valueProp || !newValue) {
@@ -240,7 +242,7 @@
       B.triggerEvent('OnChange');
     };
 
-    const getDefaultValue = React.useCallback(() => {
+    const getRecords = React.useCallback(() => {
       if (!currentValue || !results) {
         return multiple ? [] : null;
       }
@@ -263,16 +265,17 @@
       }, []);
 
       const singleRecord = currentRecords[0] ? { ...currentRecords[0] } : null;
-      return multiple ? currentRecords : singleRecord;
-    }, [results]);
 
-    const defaultRecord = getDefaultValue();
+      return multiple ? currentRecords : singleRecord;
+    }, [currentValue, results]);
+
+    const record = getRecords();
 
     useEffect(() => {
-      if (!multiple && defaultRecord && searchProp) {
-        setCurrentLabel(defaultRecord[searchProp.name]);
+      if (!multiple && record && searchProp) {
+        setCurrentLabel(record[searchProp.name]);
       }
-    }, [defaultRecord]);
+    }, [record]);
 
     const renderLabel = option => {
       const optionLabel = option[searchProp.name];
@@ -362,7 +365,7 @@
     }
 
     if (err && displayError) return <span>{err.message}</span>;
-    if (!data || hasNoProp) {
+    if (!results || hasNoProp) {
       return (
         <TextField
           {...textFieldProps}
@@ -379,9 +382,9 @@
       <Autocomplete
         multiple={multiple}
         freeSolo={freeSolo}
-        autoSelect={freeSolo}
         options={results}
-        defaultValue={defaultRecord}
+        value={record}
+        inputValue={searchParam}
         getOptionLabel={renderLabel}
         getOptionSelected={(option, value) => value.id === option.id}
         PopoverProps={{
@@ -389,8 +392,8 @@
             root: classes.popover,
           },
         }}
-        onInputChange={(_, inputValue) => {
-          setSearchParam(inputValue);
+        onInputChange={(event, inputValue) => {
+          if (event) setSearchParam(inputValue);
         }}
         onChange={onChange}
         disableCloseOnSelect={!closeOnSelect}