+++ src/components/column.js
@@ -44,7 +44,8 @@
     return ColumnComponent;
   })(),
   styles: B => theme => {
-    const { mediaMinWidth, Styling } = B;
+    const { env, mediaMinWidth, Styling } = B;
+    const isDev = env === 'dev';
     const style = new Styling(theme);
     const getSpacing = (idx, device = 'Mobile') =>
       idx === '0' ? '0rem' : style.getSpacing(idx, device);
@@ -120,7 +121,7 @@
         borderColor: 'transparent',
         borderStyle: 'none',
         borderRadius: 0,
-        overflow: 'auto',
+        overflow: isDev ? 'unset' : 'auto',
         boxSizing: 'border-box',
         [`@media ${mediaMinWidth(600)}`]: {
           display: ({