+++ src/prefabs/createForm.js
@@ -34,17 +34,21 @@
       <>
         <Header onClose={close} title="Configure form fields" />
         <Content>
-          <Field label="Select model">
+          <Field
+            label="Select model"
+            error={
+              showValidation && <Text color="#e82600">Model is required</Text>
+            }
+          >
             <ModelSelector
               onChange={id => {
                 setShowValidation(false);
                 setModelId(id);
               }}
               value={modelId}
-              margin
             />
           </Field>
-          {showValidation && <Text color="#e82600">Model is required</Text>}
+
           <Field label="Select properties">
             <PropertiesSelector
               onChange={value => {
@@ -52,6 +56,7 @@
               }}
               value={properties}
               modelId={modelId}
+              scopedModels={false}
               disabledNames={['created_at', 'updated_at', 'id']}
               disabledKinds={[
                 'BELONGS_TO',
@@ -88,7 +93,7 @@
                 .slice(1, str.length)
                 .replace(/[A-Z]/g, letter => `_${letter.toLowerCase()}`);
 
-            if (!modelId && !data && !data.model) {
+            if (!modelId || !data || !data.model) {
               setShowValidation(true);
               return;
             }
@@ -14721,6 +14726,122 @@
                   key: 'disabled',
                   value: false,
                   type: 'TOGGLE',
+                },
+                {
+                  label: 'Add Tooltip',
+                  key: 'addTooltip',
+                  value: false,
+                  type: 'TOGGLE',
+                  configuration: {
+                    as: 'VISIBILITY',
+                  },
+                },
+                {
+                  label: 'Toggle tooltip visibility',
+                  key: 'hasVisibleTooltip',
+                  value: true,
+                  type: 'TOGGLE',
+                  configuration: {
+                    as: 'VISIBILITY',
+                    condition: {
+                      type: 'SHOW',
+                      option: 'addTooltip',
+                      comparator: 'EQ',
+                      value: true,
+                    },
+                  },
+                },
+                {
+                  type: 'VARIABLE',
+                  label: 'Tooltip Content',
+                  key: 'tooltipContent',
+                  value: ['Tips'],
+                  configuration: {
+                    condition: {
+                      type: 'SHOW',
+                      option: 'addTooltip',
+                      comparator: 'EQ',
+                      value: true,
+                    },
+                  },
+                },
+                {
+                  label: 'Tooltip Placement',
+                  key: 'tooltipPlacement',
+                  value: 'bottom',
+                  type: 'CUSTOM',
+                  configuration: {
+                    as: 'DROPDOWN',
+                    dataType: 'string',
+                    allowedInput: [
+                      {
+                        name: 'Top Start',
+                        value: 'top-start',
+                      },
+                      {
+                        name: 'Top',
+                        value: 'top',
+                      },
+                      {
+                        name: 'Top End',
+                        value: 'top-end',
+                      },
+                      {
+                        name: 'Right',
+                        value: 'right',
+                      },
+                      {
+                        name: 'Left',
+                        value: 'left',
+                      },
+                      {
+                        name: 'Botttom Start',
+                        value: 'bottom-start',
+                      },
+                      {
+                        name: 'Bottom',
+                        value: 'bottom',
+                      },
+                      {
+                        name: 'Bottom End',
+                        value: 'bottom-end',
+                      },
+                    ],
+                    condition: {
+                      type: 'SHOW',
+                      option: 'addTooltip',
+                      comparator: 'EQ',
+                      value: true,
+                    },
+                  },
+                },
+                {
+                  type: 'COLOR',
+                  label: 'Tooltip Background',
+                  key: 'tooltipBackground',
+                  value: 'Medium',
+                  configuration: {
+                    condition: {
+                      type: 'SHOW',
+                      option: 'addTooltip',
+                      comparator: 'EQ',
+                      value: true,
+                    },
+                  },
+                },
+                {
+                  type: 'COLOR',
+                  label: 'Tooltip Text',
+                  key: 'tooltipText',
+                  value: 'Black',
+                  configuration: {
+                    condition: {
+                      type: 'SHOW',
+                      option: 'addTooltip',
+                      comparator: 'EQ',
+                      value: true,
+                    },
+                  },
                 },
               ],
               descendants: [],