+++ src/prefabs/media.js
@@ -27,6 +27,7 @@
           key: 'imageSource',
           type: 'VARIABLE',
           configuration: {
+            as: 'MULTILINE',
             condition: {
               type: 'SHOW',
               option: 'type',
@@ -41,6 +42,7 @@
           key: 'videoSource',
           type: 'VARIABLE',
           configuration: {
+            as: 'MULTILINE',
             condition: {
               type: 'SHOW',
               option: 'type',
@@ -55,6 +57,7 @@
           key: 'iframeSource',
           type: 'VARIABLE',
           configuration: {
+            as: 'MULTILINE',
             condition: {
               type: 'SHOW',
               option: 'type',