+++ src/prefabs/pageWithLoginForm.js
@@ -282,7 +282,7 @@
                         {
                           label: 'Size',
                           key: 'size',
-                          value: 'medium',
+                          value: 'Medium',
                           type: 'CUSTOM',
                           configuration: {
                             as: 'BUTTONGROUP',
@@ -7086,12 +7086,6 @@
               key: 'innerSpacing',
               type: 'SIZES',
             },
-            {
-              type: 'TOGGLE',
-              label: 'Allow overflow',
-              key: 'overflow',
-              value: false,
-            },
           ],
           descendants: [
             {
@@ -7566,7 +7560,7 @@
                           type: 'VARIABLE',
                           label: 'Button text',
                           key: 'buttonText',
-                          value: ['Send'],
+                          value: ['Login'],
                         },
                         {
                           type: 'CUSTOM',
@@ -8915,6 +8909,122 @@
                           key: 'disabled',
                           value: false,
                           type: 'TOGGLE',
+                        },
+                        {
+                          label: 'Add Tooltip',
+                          key: 'addTooltip',
+                          value: false,
+                          type: 'TOGGLE',
+                          configuration: {
+                            as: 'VISIBILITY',
+                          },
+                        },
+                        {
+                          label: 'Toggle tooltip visibility',
+                          key: 'hasVisibleTooltip',
+                          value: true,
+                          type: 'TOGGLE',
+                          configuration: {
+                            as: 'VISIBILITY',
+                            condition: {
+                              type: 'SHOW',
+                              option: 'addTooltip',
+                              comparator: 'EQ',
+                              value: true,
+                            },
+                          },
+                        },
+                        {
+                          type: 'VARIABLE',
+                          label: 'Tooltip Content',
+                          key: 'tooltipContent',
+                          value: ['Tips'],
+                          configuration: {
+                            condition: {
+                              type: 'SHOW',
+                              option: 'addTooltip',
+                              comparator: 'EQ',
+                              value: true,
+                            },
+                          },
+                        },
+                        {
+                          label: 'Tooltip Placement',
+                          key: 'tooltipPlacement',
+                          value: 'bottom',
+                          type: 'CUSTOM',
+                          configuration: {
+                            as: 'DROPDOWN',
+                            dataType: 'string',
+                            allowedInput: [
+                              {
+                                name: 'Top Start',
+                                value: 'top-start',
+                              },
+                              {
+                                name: 'Top',
+                                value: 'top',
+                              },
+                              {
+                                name: 'Top End',
+                                value: 'top-end',
+                              },
+                              {
+                                name: 'Right',
+                                value: 'right',
+                              },
+                              {
+                                name: 'Left',
+                                value: 'left',
+                              },
+                              {
+                                name: 'Botttom Start',
+                                value: 'bottom-start',
+                              },
+                              {
+                                name: 'Bottom',
+                                value: 'bottom',
+                              },
+                              {
+                                name: 'Bottom End',
+                                value: 'bottom-end',
+                              },
+                            ],
+                            condition: {
+                              type: 'SHOW',
+                              option: 'addTooltip',
+                              comparator: 'EQ',
+                              value: true,
+                            },
+                          },
+                        },
+                        {
+                          type: 'COLOR',
+                          label: 'Tooltip Background',
+                          key: 'tooltipBackground',
+                          value: 'Medium',
+                          configuration: {
+                            condition: {
+                              type: 'SHOW',
+                              option: 'addTooltip',
+                              comparator: 'EQ',
+                              value: true,
+                            },
+                          },
+                        },
+                        {
+                          type: 'COLOR',
+                          label: 'Tooltip Text',
+                          key: 'tooltipText',
+                          value: 'Black',
+                          configuration: {
+                            condition: {
+                              type: 'SHOW',
+                              option: 'addTooltip',
+                              comparator: 'EQ',
+                              value: true,
+                            },
+                          },
                         },
                       ],
                       descendants: [],